pyproject.toml
[tool.poetry]
name = "kkthumbs"
version = "0.1.0"
description = ""
authors = ["Antonio Feregrino <antonio.feregrino@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^8.0"
click = "^7.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
